organisation: Add models for organisational structure
We have a number of places on the site where we wish to record
information about who works for us and with us, and in what capacity.
This commit adds the following models:

- Person
- Unit
- UnitMembership
- Board
- BoardMembership

Allowing us to keep track of our staff, board members, and others, and
to record their relationships to the organisational groupings in
question.

In due course the advisory board and staff pages will be driven from
this data.