content: Use a static_placeholder for the footer
If we want a content object that is displayed and editable on every
page, we can use a static_placeholder to achieve that aim. Rather than
requiring that certain pages exist (and using some rather dodgy data
migrations to ensure that fact), we can now allow users to edit the
content of the page footer that is displayed on every page.

Hooray for Django-CMS and being sane.