Add support for Google Analytics
This commit adds support for adding the Universal Analytics tracking
code into every page on the site, provided that two configuration
variables are set:

- GOOGLE_ANALYTICS_TRACKING_ID
- GOOGLE_ANALYTICS_DOMAIN

If both of these are set at compile-time, then the GA tracking code will
be included in the JavaScript bundle for the page.

See also the notes in templates/general.html for an explanation of why
this code (usually inline) is compiled and deployed in the bundle.

Closes #33.