requirements: changed aldryn_search requirement to cms3-compat branch
django-cms version 3 includes backward incompatible changes in that
they removed the published boolean field for pages which was being
used by aldryn_search to see what pages should be indexed (only
published ones).

The current aldryn_search version therefore does not work for
django-cms3 but their github repo includes a branch that does.

Incidentally this also includes a fix for the bug where an
EmptyQuerySet was instantiated (something which was dropped in
Django 1.6). So by moving to the cms3-compat branch we solve two
problems, we no longer have to maintain our own fork with the
Django 1.6 bug fix and we can use django-cms version 3.