Views and templates for press releases and press mentions.
This change requires a database migration. It adds a slug to both
press releases and press mentions so that they can be linked to
independently (e.g. shared on social media). It also removes the
published boolean for press releases since we already have a
release date. Having a published boolean just means somebody will
at some point forget to make a press release public around the
release date and get angry/stressed because it wasn't released on
time.

Currently navigation happens mostly through the sidebar. This is
designed to be placed at the top of the press and show press
releases by default and recent press mentions on the side with a
link to all press mentions. That will then show recent press
releases in the sidebar with a link back to the main site.