press: only show the content of the most recent press release
Re-styling the templates for the press releases so that they show
only the most recent press release. All other should just be titles
and release dates (with links to the actual press releases).

This is not done to the press mentions because they have a different
nature but with this change the press releases in the sidebar of the
press mention page now also show the release date (and in the
sidebar for a single press release.