press: Don't display unpublished press releases!
We were accidentally displaying unpublished press releases on the press
mentions page. This fixes that, adds a test for it, and adds some minor
refactorings so that it's harder to do in future.

When referencing PressReleases in views for public consumption, one
should use the .published_objects manager unless there is a very good
reason to do otherwise.