organisation: style the working group templates
This uses the new feature blocks to show the workign groups along
with their logos. Removed the bootstrap template tag as we now only
show one per line.

The current icons used for the working groups have a specific
background which we hardcode into the style sheet. It would
have been better to have an image without a background but this
goes in as is to speed up the working group insertion.