organisation: add app hook to show country and regional groups
Attach this app to the main network page and it makes the network
groups accessible on subdomains /<country> and /<country>/<region>

This also removes the role field for network group memberships as
it turned out that was a misunderstanding. A member should have a
title (which describes that member's role, and not both).