organisation: add a plugin to show network group country flags
This adds a simple plugin which shows only country flags in
standardised dimensions. This plugin could be expaned later to
allow users to better control how to list network groups (i.e.
not only flags).

The flags come from the Open Icon Library and are in public domain
according to them. README from the original collection is included.
Flags were re-exported to png from the svg to ensure standard
dimensions. Some changes were made to the svg versions of a few
flag to allow export area to be the drawing and not the page
(needed for standardised dimensions). Some of the svgs contained
some extra cruft or longer lines outside of the page and therefore
needed the changes. The svgs added in this commit are the modified
ones.

It would also have been possible to use the flags that come with
django-countries but those versions do not fit very well with the
flat design and since we now have svgs we can better control the
size and dimensions.

Conflicts:
	foundation/organisation/cms_plugins.py