organisation: make working and network group model fields consistent
Projects use the form homepage_url (postfix _url) for URLFields. We
need to be consistent in how we name our fields to make it easier
to just use our models in whatever view/template we create.

This also removes a few of CharField null fields since CharFields
in Django are not nullable and will be saved as empty strings. This
changes slightly how the slugs work. We don't have to set them to
None (null) and just work directly from the empty strings.