homepage: Add featured projects section
This commit adds a "featured projects" section at the bottom of the
homepage. Multiple instances of the "Featured Project" plugin (ideally
six of them, for best appearance) can be added to this placeholder to
display a small featured projects gallery.

This commit also adds a `banner` field to the Project model, as the
pictures used in the "Featured Project" plugin may well need to be more
exciting than just a simple project logo (as stored in the `picture`
field).