templates: fix problem with search bar button being to high in FF
The search bar button was one pixel too high in Firefox but not so
in Chrome. Tweaking the line-height fixes this but is definitely
not the best solution. This line-height should at least be computed
somehow but this value (found with manual tweaking) makes it work
in both Firefox and Chrome.

This also sets the font-size to what it is in the rest of the
navbar.