Fix display of search box in Firefox
This appears to be a nasty rendering bug, possibly related to border-box
sizing in Firefox v27, 28, and possibly earlier. Relying on line-height
and vertical padding to size this button doesn't work, and fixing the
height in one place with selective CSS affects the display of other
buttons not matched by that CSS (!).

The simplest solution appears to be to explicitly set the heights of
buttons, site-wide.

Closes #116.
Closes #129.