organisation: chunk the project gallery into multiple rows
A clearer separation between the projects in each row is needed
since different heights might make not line the up as they should.
To do that a chunk template tag is needed.

Also since project gallery is timeless we wrap the pager (fixed
in previous commit to allow for timelessness) with a 'with'
statement that sets timeless to True.