search: Add updated_at field to SearchIndices
This allows the 'update_index' management command to only submit those
model objects modified within a specified timeframe to be reindexed. For
example,

    python manage.py update_index -a 2

will submit only those model objects updated in the past two hours to be
reindexed. This lightens the load on the search index, and reduces the
time spent updating the index.