search: Perform realtime index updates
Turning on the Haystack RealtimeSignalProcessor causes updates to the
search index to be made for models with appropriate on_save/on_delete
signals.

This increases the load associated with updating a model but I don't
foresee this causing any problems for us in the short term.

NB: As far as I can tell, django-cms pages are *not* automatically
updated by this mechanism, so we will additionally have to set up an
'update_index' cronjob.