Temporary fix for jobs iframe
Ugly hack to allow the Resumator Iframe to be displayed properly (Just a
huge value fo height)

I tried to remove the X-Frame-Options header on /about/jobs/helper to
see if that helped, but while it worked locally the header was not
removed on okfn.org (maybe there's another layer on top of the app I'm
not aware of). Actually I could not find at all where this helper is
called, and the local version calls
https://okfn.org/about/jobs/helper/?height=170&cacheb=0.973274856087319&nojump=0
which is just bizarre.