remove django-geoposition package
This dependency is problematic as we upgrade
it hasn't been updated since 2016.

Our usage of this is minimal. We store an
arbitrary point against some Network Groups.
It is not used at all on the website,
we just chuck it out in the CSV
(which was broken until this PR anyway).

Lets completely remove the one GeopositionField
If we find we need this functionality again in
future, we can revisit a solution
(PostGIS/SpatiaLite would probably be more
appropriate but also I suspect we just don't
need this functionality at all).